refactor assign method and move CreateKymaRuntime and revert some tests
