+++ tests/director/tests/token_api_test.go
@@ -26,8 +26,10 @@ func TestTokenGeneration(t *testing.T) {
 
 		input := fixRuntimeInput("test")
 
-		runtime := fixtures.RegisterKymaRuntime(t, ctx, certSecuredGraphQLClient, tenantId, input, conf.GatewayOauth)
+		runtime, err := fixtures.RegisterRuntimeFromInputWithinTenant(t, ctx, certSecuredGraphQLClient, tenantId, &input)
 		defer fixtures.CleanupRuntime(t, ctx, certSecuredGraphQLClient, tenantId, &runtime)
+		require.NoError(t, err)
+		require.NotEmpty(t, runtime.ID)
 		tokenRequestNumber := 3
 
 		//WHEN
